refactor: Set default Content-Type header in axiosConfig
The `setRequestDefaultHeader` function in `axiosConfig.ts` has been updated to set the default `Content-Type` header for Axios requests. This ensures that all requests sent by the application have the correct content type of `application/json;charset=utf-8`. The function now checks if the `headers` property exists in the request configuration and sets the header accordingly. If the `headers` property does not exist, it creates a new `AxiosHeaders` object and sets the `Content-Type` header. This refactor improves the consistency and reliability of the request headers in the application.